Task #2 moved to separate package
